SECTION 5 Identity Server Issues Issue with JWT and HTTP POST

+++ src/AuctionService/Controllers/AuctionsController.cs
@@ -5,6 +5,7 @@ using AutoMapper;
 using AutoMapper.QueryableExtensions;
 using Contracts;
 using MassTransit;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 
@@ -54,12 +55,15 @@ public class AuctionsController : ControllerBase
 
    }
 
+
+   [Authorize]
    [HttpPost]
     public async Task<ActionResult<AuctionDTO>>CreateAuction(CreateAuctionDTO auctionDTO)
     {
         var auction = _mapper.Map<Auction>(auctionDTO);
-        //TODO add current user as seller
-        auction.Seller="test";
+        
+        
+        auction.Seller= User.Identity.Name;
 
         _context.Auctions.Add(auction);
         
@@ -75,6 +79,7 @@ public class AuctionsController : ControllerBase
         new {auction.Id}, _mapper.Map<AuctionDTO>(auction));
     }
 
+    [Authorize]
     [HttpPut("{id}")]
 
     public async Task<ActionResult> UpdateAuction(Guid id, UpdateAuctionDTO updateAuctionDTO)
@@ -82,9 +87,11 @@ public class AuctionsController : ControllerBase
         var auction= await _context.Auctions.Include(x => x.Item)
             .FirstOrDefaultAsync( x => x.Id == id);
 
+        if (auction.Seller != User.Identity.Name) return Forbid();    
+
         if (auction == null)  return NotFound();
 
-        //Todo : check seller == username
+        
 
         auction.Item.Make = updateAuctionDTO.Make ?? auction.Item.Make;
         auction.Item.Model = updateAuctionDTO.Model ?? auction.Item.Model;
@@ -103,6 +110,7 @@ public class AuctionsController : ControllerBase
 
     }
 
+    [Authorize]
     [HttpDelete("{id}")]
 
     public async Task<ActionResult>DeleteAuction(Guid id)
@@ -111,8 +119,7 @@ public class AuctionsController : ControllerBase
 
         if (auction == null) return NotFound();
 
-        //Todo : check seller === usewrname
-
+        if (auction.Seller != User.Identity.Name) return Forbid();
         _context.Auctions.Remove(auction);
 
         await _publishEndpoint.Publish<AuctionDeleted>(new{Id = auction.Id.ToString()});
